* Changed STDDEV formula from population to sample * Fixed Normalization issues (maybe, further inquiry required) * Added Min and Max FT vectors to configuration file * Now by default train and test function will use the program config instead of needing to specify one (but you can still do so) * Added few utility functions to accelerometer/features.h

+++ src/accelerometer/accelerometer.cpp
@@ -29,13 +29,12 @@ namespace phd {
             Acceleration Accelerometer::fetch() {
                 NunchuckData values = dataStore->fetch();
                 Acceleration result = {
-                        values.getAccelerationValues().X,
-                        values.getAccelerationValues().Y,
-                        values.getAccelerationValues().Z
+                        static_cast<float>(values.getAccelerationValues().X),
+                        static_cast<float>(values.getAccelerationValues().Y),
+                        static_cast<float>(values.getAccelerationValues().Z)
                 };
                 return result;
             }
-
 
             SimulatedAccelerometer::SimulatedAccelerometer(const std::string &path_to_acceleration_file){
                 this->currentElement = 0;
